Refactor MapFilterContext
