+++ src/components/map/ObservationDateOverlay.tsx
@@ -2,8 +2,7 @@ import { DateTime } from 'luxon';
 import { useContext } from 'react';
 import { useTranslation } from 'react-i18next';
 import { MdLayers } from 'react-icons/md';
-import { setObservationDate } from '../../context/MapFilterActions';
-import { MapFilterContext } from '../../context/MapFilterContext';
+import { MapFilterContext, setObservationDate } from '../../context/MapFilter';
 import { useFilterStops } from '../../hooks';
 import { Column, Row } from '../../layoutComponents';
 import { IconButton } from '../../uiComponents';
